# code/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type KeepaliveStatusDatabase = typeof KeepaliveStatusDatabase[keyof typeof KeepaliveStatusDatabase];


export const KeepaliveStatusDatabase = {
  ok: 'ok',
  cold: 'cold',
  not_configured: 'not_configured',
} as const;

export interface KeepaliveStatus {
  status: string;
  database: KeepaliveStatusDatabase;
  timestamp: string;
}

/**
 * Which Vedika environment produced this data
 */
export type PanchangSource = typeof PanchangSource[keyof typeof PanchangSource];


export const PanchangSource = {
  sandbox: 'sandbox',
  production: 'production',
} as const;

export type PanchangTithi = {
  name: string;
  paksha: string;
  number: number;
};

export type PanchangNakshatra = {
  name: string;
  lord: string;
};

export type PanchangMasa = {
  name: string;
  englishEquivalent: string;
};

export type PanchangDishaShool = {
  direction: string;
  description: string;
};

/**
 * Simplified Panchang summary derived from the Vedika API
 */
export interface Panchang {
  /** Which Vedika environment produced this data */
  source: PanchangSource;
  date: string;
  tithi: PanchangTithi;
  nakshatra: PanchangNakshatra;
  masa: PanchangMasa;
  overallAuspiciousness: string;
  summary: string;
  bestActivities: string[];
  activitiesToAvoid: string[];
  dishaShool: PanchangDishaShool;
}

export interface ApiErrorMessage {
  error: string;
}

/**
 * Body for sending a free-form WhatsApp text message
 */
export interface WhatsappMessageInput {
  /** Recipient phone number in E.164 format (e.g. +919876543210) */
  to: string;
  message: string;
}

export type WhatsappSendResultStatus = typeof WhatsappSendResultStatus[keyof typeof WhatsappSendResultStatus];


export const WhatsappSendResultStatus = {
  sent: 'sent',
} as const;

export interface WhatsappSendResult {
  status: WhatsappSendResultStatus;
  messageId: string;
}

export interface WhatsappInboundMessage {
  from: string;
  text: string;
  receivedAt: string;
}

export interface WhatsappOutboundMessage {
  to: string;
  text: string;
  sentAt: string;
}

export interface PurohitRecord {
  id: string;
  phoneNumber: string;
  name: string;
  city: string;
  latitude: number;
  longitude: number;
  localityKey: string;
  upiId: string;
  calendarSystem: string;
  plan: string;
  createdAt: string;
}

export type IngestJobStatusKind = typeof IngestJobStatusKind[keyof typeof IngestJobStatusKind];


export const IngestJobStatusKind = {
  voice: 'voice',
  photo: 'photo',
} as const;

export type IngestJobStatusStatus = typeof IngestJobStatusStatus[keyof typeof IngestJobStatusStatus];


export const IngestJobStatusStatus = {
  received: 'received',
  transcribed: 'transcribed',
  extracted: 'extracted',
  awaiting_confirm: 'awaiting_confirm',
  confirmed: 'confirmed',
  rejected: 'rejected',
  failed: 'failed',
} as const;

export type IngestJobStatusFieldScores = {[key: string]: number} | null;

export interface IngestJobStatus {
  id: string;
  purohitId: string;
  kind: IngestJobStatusKind;
  status: IngestJobStatusStatus;
  fieldScores?: IngestJobStatusFieldScores;
  error?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface PurgeResult {
  purgedTranscripts: number;
  purgedExtractions: number;
  expiredAwaitingConfirm: number;
  deletedRows: number;
}

export type LedgerStatusPaymentStatus = typeof LedgerStatusPaymentStatus[keyof typeof LedgerStatusPaymentStatus];


export const LedgerStatusPaymentStatus = {
  pending: 'pending',
  claimed: 'claimed',
  corroborated: 'corroborated',
} as const;

export interface LedgerStatus {
  id: string;
  purohitId: string;
  yajmanId: string;
  eventId?: string | null;
  amountCollected?: string | number | null;
  paymentStatus: LedgerStatusPaymentStatus;
  purohitClaimedAt?: string | null;
  familyConfirmedAt?: string | null;
  localityKey: string;
  createdAt: string;
}

export interface CohortMetrics {
  week: string;
  cohortSize: number;
  referredActivations: number;
  observedK: number;
}

export type CohortMetricsList = CohortMetrics[];

export type GetPanchangParams = {
/**
 * Date to resolve, ISO 8601 (defaults to today)
 */
date?: string;
latitude?: number;
longitude?: number;
};