# code/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /purohits/{phoneNumber}:
    get:
      operationId: getPurohitByPhone
      tags: [purohits]
      summary: Get a purohit's persisted onboarding record by phone number
      parameters:
        - name: phoneNumber
          in: path
          required: true
          schema:
            type: string
        - name: X-Internal-Key
          in: header
          required: true
          description: Shared secret gating this PII-exposing inspection endpoint -- must match INTERNAL_API_KEY
          schema:
            type: string
      responses:
        "200":
          description: Purohit onboarding record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurohitRecord"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
        "404":
          description: Purohit not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
...
components:
  schemas:
...
    ApiErrorMessage:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    PurohitRecord:
      type: object
      properties:
        id:
          type: string
        phoneNumber:
          type: string
        name:
          type: string
        city:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        localityKey:
          type: string
        upiId:
          type: string
        calendarSystem:
          type: string
        plan:
          type: string
        createdAt:
          type: string
      required:
        - id
        - phoneNumber
        - name
        - city
        - latitude
        - longitude
        - localityKey
        - upiId
        - calendarSystem
        - plan
        - createdAt
...