# code/lib/api-zod/src/generated/types/purohitRecord.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PurohitRecord {
  id: string;
  phoneNumber: string;
  name: string;
  city: string;
  latitude: number;
  longitude: number;
  localityKey: string;
  upiId: string;
  calendarSystem: string;
  plan: string;
  createdAt: string;
}